Also rename the test module
