+++ parley_tests/tests/mod.rs
@@ -23,7 +23,6 @@
 #![allow(clippy::cast_possible_truncation, reason = "not critical for testing")]
 
 mod basic;
-mod chrome_linebreaking;
 mod cursor;
 mod draw;
 mod editor;
@@ -31,6 +30,7 @@ mod exclusions;
 mod floats;
 mod issues;
 mod line_break;
+mod linebreaking_matches_chrome;
 mod lines;
 mod out_of_flow_boxes;
 mod styles;
